requirement.txt
google-generativeai
python-dotenv
pandas
openpyxl
